Bot structure was changed * add bootstrap file for loading bot scripts through robot API * main script was moved into src path
